Create landings subject folder (#37416)
