+++ components/sidebar/SidebarNav.tsx
@@ -1,8 +1,8 @@
 import cx from 'classnames'
 
 import { useMainContext } from 'components/context/MainContext'
-import { SidebarProduct } from './SidebarProduct'
-import { SidebarHomepage } from './SidebarHomepage'
+import { SidebarProduct } from 'src/landings/components/SidebarProduct'
+import { SidebarHomepage } from '../../src/landings/components/SidebarHomepage'
 import { AllProductsLink } from './AllProductsLink'
 import { ApiVersionPicker } from 'src/rest/components/ApiVersionPicker'
 import { Link } from 'components/Link'
