+++ middleware/index.js
@@ -50,7 +50,7 @@ import glossaries from './contextualizers/glossaries.js'
 import features from './contextualizers/features.js'
 import productExamples from './contextualizers/product-examples.js'
 import productGroups from './contextualizers/product-groups.js'
-import featuredLinks from './featured-links.js'
+import featuredLinks from '../src/landings/middleware/featured-links.js'
 import learningTrack from './learning-track.js'
 import next from './next.js'
 import renderPage from './render-page.js'
