+++ src/landings/README.md
@@ -1,0 +1,7 @@
+# Landing pages
+
+Landing pages include the Docs home page, the product landing pages, and product guides pages.
+
+## What landing pages are for
+
+Landing pages provide a hierarchical view of their area, making navigating the Docs easier.
