+++ src/landings/components/ArticleList.tsx
@@ -4,7 +4,7 @@ import { ActionList } from '@primer/react'
 import { useTranslation } from 'components/hooks/useTranslation'
 import { Link } from 'components/Link'
 import { ArrowRightIcon } from '@primer/octicons-react'
-import { FeaturedLink } from 'components/context/ProductLandingContext'
+import { FeaturedLink } from 'src/landings/components/ProductLandingContext'
 import { useMainContext } from 'components/context/MainContext'
 import { TruncateLines } from 'components/ui/TruncateLines'
 import { BumpLink } from 'components/ui/BumpLink'
