+++ src/landings/components/ProductGuides.tsx
@@ -1,9 +1,9 @@
 import { DefaultLayout } from 'components/DefaultLayout'
-import { useProductGuidesContext } from 'components/context/ProductGuidesContext'
-import { LandingSection } from 'components/landing/LandingSection'
-import { GuidesHero } from 'components/guides/GuidesHero'
-import { LearningTracks } from 'components/guides/LearningTracks'
-import { ArticleCards } from 'components/guides/ArticleCards'
+import { useProductGuidesContext } from 'src/landings/components/ProductGuidesContext'
+import { LandingSection } from 'src/landings/components/LandingSection'
+import { GuidesHero } from 'src/landings/components/GuidesHero'
+import { LearningTracks } from 'src/landings/components/LearningTracks'
+import { ArticleCards } from 'src/landings/components/ArticleCards'
 import { useTranslation } from 'components/hooks/useTranslation'
 import { useMainContext } from 'components/context/MainContext'
