+++ src/landings/components/ProductLanding.tsx
@@ -1,16 +1,16 @@
 import { useRouter } from 'next/router'
 import { DefaultLayout } from 'components/DefaultLayout'
-import { useProductLandingContext } from 'components/context/ProductLandingContext'
+import { useProductLandingContext } from 'src/landings/components/ProductLandingContext'
 
-import { LandingHero } from 'components/landing/LandingHero'
-import { FeaturedArticles } from 'components/landing/FeaturedArticles'
-import { GuideCards } from 'components/landing/GuideCards'
-import { SponsorsExamples } from 'components/landing/SponsorsExamples'
-import { CommunityExamples } from 'components/landing/CommunityExamples'
-import { LandingSection } from 'components/landing/LandingSection'
+import { LandingHero } from 'src/landings/components/LandingHero'
+import { FeaturedArticles } from 'src/landings/components/FeaturedArticles'
+import { GuideCards } from 'src/landings/components/GuideCards'
+import { SponsorsExamples } from 'src/landings/components/SponsorsExamples'
+import { CommunityExamples } from 'src/landings/components/CommunityExamples'
+import { LandingSection } from 'src/landings/components/LandingSection'
 import { useTranslation } from 'components/hooks/useTranslation'
-import { ProductArticlesList } from 'components/landing/ProductArticlesList'
-import { ProductReleases } from 'components/landing/ProductReleases'
+import { ProductArticlesList } from 'src/landings/components/ProductArticlesList'
+import { ProductReleases } from 'src/landings/components/ProductReleases'
 import { useVersion } from 'components/hooks/useVersion'
 import { RestRedirect } from 'src/rest/components/RestRedirect'
