+++ src/landings/components/ProductReleases.tsx
@@ -1,6 +1,6 @@
 import { ArrowRightIcon, ArrowUpIcon, FileIcon, ListUnorderedIcon } from '@primer/octicons-react'
 import { useMainContext } from 'components/context/MainContext'
-import { useProductLandingContext } from 'components/context/ProductLandingContext'
+import { useProductLandingContext } from 'src/landings/components/ProductLandingContext'
 import { useTranslation } from 'components/hooks/useTranslation'
 import { Link } from 'components/Link'
 import { useRouter } from 'next/router'
