+++ src/landings/components/TocLanding.tsx
@@ -4,10 +4,10 @@ import cx from 'classnames'
 import { useTocLandingContext } from 'components/context/TocLandingContext'
 import { useTranslation } from 'components/hooks/useTranslation'
 import { DefaultLayout } from 'components/DefaultLayout'
-import { TableOfContents } from 'components/landing/TableOfContents'
+import { TableOfContents } from 'src/landings/components/TableOfContents'
 import { ArticleTitle } from 'components/article/ArticleTitle'
 import { MarkdownContent } from 'components/ui/MarkdownContent'
-import { ArticleList } from 'components/landing/ArticleList'
+import { ArticleList } from 'src/landings/components/ArticleList'
 import { ArticleGridLayout } from 'components/article/ArticleGridLayout'
 import { Callout } from 'components/ui/Callout'
 import { Lead } from 'components/ui/Lead'
