+++ src/landings/middleware/featured-links.js
@@ -1,5 +1,5 @@
-import getLinkData from '../lib/get-link-data.js'
-import renderContent from '../lib/render-content/index.js'
+import getLinkData from '../../../lib/get-link-data.js'
+import renderContent from '../../../lib/render-content/index.js'
 
 // this middleware adds properties to the context object
 export default async function featuredLinks(req, res, next) {
